requirements.txt
google-api-python-client==2.141.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
